Second Commit: Added Second Task with tests changed Map<String,Float> to Map<String,BigDecimal> in FirstTask

+++ src/main/java/org/models/Fine.java
@@ -3,18 +3,20 @@ package org.models;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 @JsonAutoDetect
 public class Fine {
     @JsonProperty("fine_amount")
-    private Float fineAmount;
+    private BigDecimal fineAmount;
     @JsonProperty("type")
     private String type;
 
-    public Float getFineAmount() {
+    public BigDecimal getFineAmount() {
         return fineAmount;
     }
 
-    public void setFineAmount(Float fineAmount) {
+    public void  setFineAmount(BigDecimal fineAmount) {
         this.fineAmount = fineAmount;
     }
